Trying with coalesced partitions

+++ tweetset_loader.py
@@ -371,7 +371,7 @@ if __name__ == '__main__':
                                 path_to_dataset=filepath_list,
                                 dataset_id=dataset_id)
             # Reduce number of partitions to constrain number of gzipped files creates
-            num_partitions = (tweet_count / 25000) or 1
+            num_partitions = int(tweet_count / 25000) or 1
             df = df.coalesce(num_partitions)
             # Create and save tweet ID's
             tweet_ids_path = os.path.join(full_dataset_path, 'tweet-ids')
